% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bin_quantiles}
\alias{bin_quantiles}
\title{Split values into discrete, equal sized bins.}
\usage{
bin_quantiles(x, n, num_labels = FALSE)
}
\arguments{
\item{x}{Values to bin.}

\item{n}{Number of bins.}

\item{num_labels}{Label the bins with simple sequence of numbers?}
}
\value{
Factor of bins.
}
\description{
Split values into discrete, equal sized bins.
}
