% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Complementary function to \%in\%.}
\usage{
x \%notin\% vec
}
\arguments{
\item{x}{Vector of elements to be matched.}

\item{vec}{Vector of elements to be matched against.}
}
\value{
Logical vector of the same length as x.
}
\description{
Complementary function to \%in\%.
}
