% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%+\%}
\alias{\%+\%}
\title{Paste two strings using an infix operator.}
\usage{
a \%+\% b
}
\arguments{
\item{a}{Character string.}

\item{b}{Character string.

 @return Character string as paste0(a, b).}
}
\description{
Paste two strings using an infix operator.
}
