% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_dim}
\alias{set_dim}
\title{Set dimensions and resolution of a plot in an R kernel-based Jupyter
notebook.}
\usage{
set_dim(width, height, res = 300)
}
\arguments{
\item{width}{Width in inches.}

\item{height}{Height in inches.}

\item{res}{PPI for rasterization.}
}
\description{
Set dimensions and resolution of a plot in an R kernel-based Jupyter
notebook.
}
